--- app/Controllers/AboutController.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Controllers;
-
-use App\Core\Controller;
-use App\Core\Template;
-
-class AboutController extends Controller
-{
-    public function __invoke(): Template
-    {
-        return view("about", [
-            'title' => 'About Us',
-        ]);
-    }
-}
