+++ app/HTTP/Controllers/AboutController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace app\HTTP\Controllers;
+
+use App\Core\Controller;
+use App\Core\Template;
+
+class AboutController extends Controller
+{
+    public function __invoke(): Template
+    {
+        return view("about", [
+            'title' => 'About Us',
+        ]);
+    }
+}
