+++ app/Services/CategoryService.php
@@ -3,8 +3,8 @@
 namespace App\Services;
 
 use App\Core\ServiceProvider;
+use app\Enums\CategoryStatus;
 use App\Models\Category;
-use App\Enums\CategoryStatus;
 
 class CategoryService extends ServiceProvider
 {
