--- app/Traits/EnumMethods.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace App\Traits;
-
-trait EnumMethods
-{
-    public static function toValues(): array
-    {
-        // use the cases() method to return the values of the enum in an array dynamically
-        $values = [];
-        foreach (self::cases() as $case) {
-            $values[] = $case->value;
-        }
-        return $values;
-    }
-
-}
